backend/requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
sqlmodel
python-jose[cryptography]
argon2-cffi
psycopg2-binary
passlib
python-dotenv
requests
python-multipart
